migrate chart stuff to its own tile control.
