+++ WalletWasabi.Fluent/ViewModels/Wallets/WalletViewModel.cs
@@ -17,26 +17,11 @@ using WalletWasabi.Wallets;
 
 namespace WalletWasabi.Fluent.ViewModels.Wallets
 {
-	public readonly struct DataLegend
-	{
-		public DataLegend(Money amount, string label, string hexColor, double percentShare)
-		{
-			Amount = amount;
-			Label = label;
-			HexColor = hexColor;
-			PercentShare = percentShare;
-		}
 
-		public Money Amount { get; }
-		public string Label { get; }
-		public string HexColor { get; }
-		public double PercentShare { get; }
-	}
 
 	public partial class WalletViewModel : WalletViewModelBase
 	{
-		[AutoNotify] private IList<(string color, double percentShare)> _testDataPoints;
-		[AutoNotify] private IList<DataLegend> _testDataPointsLegend;
+
 
 		protected WalletViewModel(UiConfig uiConfig, Wallet wallet) : base(wallet)
 		{
@@ -65,19 +50,9 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets
 				})
 				.DisposeWith(Disposables);
 
-			TestDataPoints = new List<(string, double)>()
-			{
-				("#72BD81", 0.8d),
-				("#F9DE7D", 0.2d)
-			};
-
-			TestDataPointsLegend = new List<DataLegend>
-			{
-				new (Money.Parse("0.77508"),"Private", "#F9DE7D", 0.2 ),
-				new (Money.Parse("3.10032"),"Not Private", "#72BD81", 0.8)
-			};
 			History = new HistoryViewModel(wallet);
 			BalanceTile = new WalletBalanceTileViewModel(wallet);
+			WalletPieChart = new WalletPieChartTileViewModel(wallet);
 		}
 
 		private CompositeDisposable Disposables { get; set; }
@@ -87,6 +62,8 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets
 		public HistoryViewModel History { get; }
 
 		public WalletBalanceTileViewModel BalanceTile { get; }
+
+		public WalletPieChartTileViewModel WalletPieChart { get; }
 
 		protected override void OnNavigatedTo(bool isInHistory, CompositeDisposable disposables)
 		{
