+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/WalletPieChartTileView.axaml
@@ -1,0 +1,48 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:controls="clr-namespace:WalletWasabi.Fluent.Controls"
+             xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles"
+             xmlns:converters="clr-namespace:WalletWasabi.Fluent.Converters"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletPieChartTileView">
+    <Border ClipToBounds="False" CornerRadius="5" Background="#2b2b2b" Margin="6" Padding="16"
+            BoxShadow="{DynamicResource InfoBoxShadow}">
+        <DockPanel>
+            <TextBlock Text="Wallet Privacy Status" DockPanel.Dock="Top" />
+            <DockPanel LastChildFill="True" TextBlock.FontSize="11">
+
+                <StackPanel DockPanel.Dock="Top">
+                    <TextBlock>Coin Privacy</TextBlock>
+                    <TextBlock Opacity="0.5">in Percentage %</TextBlock>
+                </StackPanel>
+
+                <ItemsRepeater DockPanel.Dock="Bottom" Margin="0,0,0,-8" Items="{Binding TestDataPointsLegend}">
+                    <ItemsRepeater.ItemTemplate>
+                        <DataTemplate x:DataType="wallets:DataLegend">
+                            <DockPanel Margin="0,8">
+                                <Ellipse DockPanel.Dock="Left" Margin="0,0,8,0" HorizontalAlignment="Left" Width="10"
+                                         Height="10"
+                                         Fill="{Binding HexColor, Converter={x:Static converters:ColorStringConverters.HexColorToBrush}}" />
+                                <TextBlock DockPanel.Dock="Left" HorizontalAlignment="Left"
+                                           Text="{Binding Amount, Converter={x:Static converters:MoneyConverters.MoneyToString}}" />
+                                <TextBlock DockPanel.Dock="Right" Margin="8,0,0,0" HorizontalAlignment="Right"
+                                           Text="{Binding PercentShare, StringFormat=\{0:P1\}}" />
+                                <TextBlock DockPanel.Dock="Right" HorizontalAlignment="Right" Text="{Binding Label}" />
+                            </DockPanel>
+                        </DataTemplate>
+                    </ItemsRepeater.ItemTemplate>
+                    <ItemsRepeater.Layout>
+                        <StackLayout />
+                    </ItemsRepeater.Layout>
+                </ItemsRepeater>
+
+                <Panel Margin="8" VerticalAlignment="Center" HorizontalAlignment="Center">
+                    <controls:RingChartControl DataPoints="{Binding TestDataPoints}" />
+                </Panel>
+            </DockPanel>
+        </DockPanel>
+    </Border>
+
+</UserControl>
