+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/WalletPieChartTileView.axaml.cs
@@ -1,0 +1,19 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Fluent.Views.Wallets.Home.Tiles
+{
+	public class WalletPieChartTileView : UserControl
+	{
+		public WalletPieChartTileView()
+		{
+			InitializeComponent();
+		}
+
+		private void InitializeComponent()
+		{
+			AvaloniaXamlLoader.Load(this);
+		}
+	}
+}
